+++ src/Teams.Notifications.AdaptiveCardGen/AdaptiveCardTemplateGenerator.cs
@@ -48,11 +48,11 @@ public class AdaptiveCardTemplateGenerator : IIncrementalGenerator
         var modelProperties = content.GetMustachePropertiesFromString();
         var modelName = $"{fileName}Model";
         var controllerName = $"{fileName}Controller";
-    
+
         var modelSource = GenerateModel(modelName, modelProperties);
         spc.AddSource($"{modelName}.g.cs", SourceText.From(modelSource, Encoding.UTF8));
 
-        var controllerSource = GenerateController(modelName, controllerName, fileName, spc);
+        var controllerSource = GenerateController(fileName, spc);
         spc.AddSource($"{controllerName}.g.cs", SourceText.From(controllerSource, Encoding.UTF8));
     }
 
@@ -121,7 +121,7 @@ public class AdaptiveCardTemplateGenerator : IIncrementalGenerator
               """;
     }
 
-    private static string GenerateController(string modelName, string controllerName, string name, SourceProductionContext spc)
+    private static string GenerateController(string name, SourceProductionContext spc)
     {
         var text = ReadResource("CardTemplateController.csgen", spc)
             .Replace("{{name}}", name);
