+++ src/Teams.Notifications.Api/GlobalUsings.cs
@@ -21,13 +21,11 @@ global using AdaptiveCards;
 global using Azure.Core;
 global using Azure.Identity;
 global using Microsoft.Agents.Authentication;
-global using Microsoft.Agents.Authentication.Msal;
 global using Microsoft.Agents.Builder;
 global using Microsoft.Agents.Builder.App;
 global using Microsoft.Agents.Builder.State;
 global using Microsoft.Agents.Core.Models;
 global using Microsoft.Agents.Core.Serialization;
-global using Microsoft.Agents.Extensions.Teams;
 global using Microsoft.Agents.Hosting.AspNetCore;
 global using Microsoft.Agents.Storage;
 global using Microsoft.AspNetCore.Authentication.JwtBearer;
@@ -36,7 +34,6 @@ global using Microsoft.AspNetCore.Builder;
 global using Microsoft.AspNetCore.Http;
 global using Microsoft.AspNetCore.Mvc;
 global using Microsoft.AspNetCore.Mvc.ApplicationModels;
-global using Microsoft.AspNetCore.Mvc.Filters;
 global using Microsoft.Extensions.Configuration;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.Logging;
@@ -52,7 +49,6 @@ global using Microsoft.Kiota.Abstractions;
 global using Microsoft.OpenApi.Models;
 global using Polly;
 global using Swashbuckle.AspNetCore.Annotations;
-global using Teams.Notifications.AdaptiveCardGen;
 global using Teams.Notifications.Api;
 global using Teams.Notifications.Api.Action.Models;
 global using Teams.Notifications.Api.Agents;
