+++ src/Teams.Notifications.Api/Teams.Notifications.Api.csproj
@@ -51,12 +51,12 @@
     <PackageReference Include="Microsoft.SemanticKernel.Connectors.AzureOpenAI" Version="*" />
     <PackageReference Include="Microsoft.SemanticKernel.Connectors.OpenAI" Version="*" />
     <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="9.*" />
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="*" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="*" />
     <PackageReference Include="Swashbuckle.AspNetCore.Annotations" Version="*" />
     <PackageReference Include="Microsoft.Extensions.Http.Polly" Version="*" />
     <PackageReference Include="Microsoft.Graph.Beta" Version="5.*-*" />
-    <ProjectReference Include="..\Teams.Notifications.AdaptiveCardGen\Teams.Notifications.AdaptiveCardGen.csproj" OutputItemType="Analyzer" />
+    <ProjectReference Include="..\Teams.Notifications.AdaptiveCardGen\Teams.Notifications.AdaptiveCardGen.csproj"
+                      OutputItemType="Analyzer" />
     <AdditionalFiles Include="Templates\**\*.json" />
   </ItemGroup>
