+++ src/Teams.Notifications.Api/Telemetry/TelemetryExtensions.cs
@@ -44,7 +44,7 @@ internal static class TelemetryExtensions
         var appInsightsConnectionString = builder.Configuration["APPLICATIONINSIGHTS:CONNECTIONSTRING"];
         builder.Logging.ClearProviders();
 
-       var openTelemetryBuilder= builder
+        var openTelemetryBuilder = builder
             .Services
             .AddOpenTelemetry()
             .ConfigureResource(r =>
@@ -82,8 +82,8 @@ internal static class TelemetryExtensions
                 .AddRuntimeInstrumentation()
                 .AddAspNetCoreInstrumentation()
             );
-       if (builder.Environment.EnvironmentName != "local")
-           openTelemetryBuilder.UseAzureMonitor(options => { options.ConnectionString = appInsightsConnectionString; });
+        if (builder.Environment.EnvironmentName != "local")
+            openTelemetryBuilder.UseAzureMonitor(options => { options.ConnectionString = appInsightsConnectionString; });
     }
 
     public static Dictionary<string, object> ToDictionary(this object obj)
